+++ NEWS.md
@@ -2,9 +2,13 @@
 
 ## Documentation improvements
 
+* New `vignette("using-credit-roles")` which describes how to work with the [Contributor Roles Taxonomy](https://credit.niso.org) in plume.
+
 * New `vignette("plume-workflow")` which describes how to use plume and [googlesheets4](https://googlesheets4.tidyverse.org) to manage author metadata.
 
 ## Minor improvements and bug fixes
+
+* `$new()` and `plm_template()` gain a new parameter `credit_roles` to facilitate the use of the [Contributor Roles Taxonomy](https://credit.niso.org).
 
 * Fixed issues when pushing metadata to empty YAML headers (#9).